habilitado cors

+++ src/main/java/br/com/petzon/petzonapi/repository/PetRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PetRepository extends JpaRepository<Pet, Long> {
+public interface PetRepository extends JpaRepository<Pet, Integer> {
 
     List<Pet> findByTipo(PetType tipo);
 }
